Register Step:
The Register Step uses the two given images and co-registers them.
Since we are using BRAINS for the registration, we need to build it with
Slicer.
The Register Step was thus fitted to use the BrainsFit CLI. So far the
options allowed are limited to the inputs as the module used might be
changed in the future.

This is also the occasion to add some infrastructure to the WorkflowStep
and the Workflow. A progress bar is added above the advanced panel.