Add initial step
The initial step is in charge of choosing what kind of analisys will be
performed in the workflow. For each of these analysis, a json file will be
added in the resources to help the CLIs.
Right now there are now json files, so Readme are added to make sure the
folder are taken into account by git.