Update workflow GUI from CLIs
The workflow was already updating the CLIs in slicer with its own
parameters. Now, the workflow GUI can update itself from the CLIs it uses
by overloading the method updateFromCLIParameters(). This method is called
by the workflow on the steps everytime the worflow module is shown. The
default implementation in WorkflowStep does nothing. So far, this only
applied to the RegisterStep.

Also, instead of using BrainsFit, the registration uses the cli
ExpertAutomatedRegistration.