Add Resample Step
The resample step is in charge of making sure the input image is isotropic
so the rest of the algorithm can work better. It uses TubeTK's resample
image filter with the option isotropic.
If the input image happens to be isotropic, the user is prompt a question
to make sure he wants to resample anyway. In that case, the step is
automatically validated