Add Editor to SegmentationStep
The editor is added as a widget here. So fare, this step has the following
properties:
The master volume is controlled by a combobox (advanced mode) and the
register step output.
OnEntry(), the master volume is set so the pop up window doesn't appear
before arriving in the module.
Upon setting the master volume, the step looks for the merge volume by
searching through the scene for a node that has a corresponding name. This
is the merge node and is selected in the combobox.
The step itself is validated only if the current merge node has more than
2 labels. This is done by verifying that the range min and the range max
of the image scalars are different.