Fix weird spacing above Advanced Settings Panel
There was a small gap between the workflow and the advanced settings panel
due to a bad top margin in the advanced settings panel ui.
Also, adding left and right spacing so it's not stuck against the borders
of the application. This is because the prometheus workflow layout doesn't
have any borders.