Infere the names from json config file
Depending on the task selected, we want to be able to change the names of
the workflow inputs dynamicaly to provide the user with a clearer GUI.
This is done in the intial step. Much like it is done for the json presets,
the naming is stored in a json file named Workflow.json.
The README were consequently updated to reflect this too.
Finaly, the naming was propagated to the LoadDataStep, the ResampleStep
and the SegmentationStep.