ENH: Improve view update for the workflow
Instead of relying on an internal map of volumes, it is much better to
use the selection node. Since it already keeps track of the volumes
displayed, we can use it to make sure we are on sync with them.
The update is now automatically called when the selection node is modified.