ENH: Support 1 to 3 inputs in Register step
In order to reduce code redundancy, the register step was separated in
a widget and a step itself. The widget is not much different than the
old step as it's in charge of running the CLI.
The step is now in charge of instantiating the necessary widgets and make
sure to validate them correctly.