ENH: Create an adaptive layout
Create a different layout when dealing with 1, 2 or 3 inputs. This layout
is update through the use of the setViews method.
The workflow keeps track of all the volume displayed to make sure that
added volumes aren't shown after a setView update.
It also take care of enabling/showing the opacity slider whenever a slice
has more than one image displayed.

WIP