ENH: Fix workflow module palette
As explained in [1], since the palette of the Slicer designer plugin base
class (see [2]) is explicitly set, the application wide palette is not
propagated. This commit workaround this issue by explicitly setting the
palette of the widget loaded from UI file.

To find out which widget have an explicit palette, the following
snipped can be used:

for w in slicer.app.allWidgets():
    if w.testAttribute(qt.Qt.WA_SetPalette):
        print("%s - [%s] (%s)" % (w, w.objectName, w.parent()))

[1] http://blog.qt.digia.com/blog/2008/11/16/font-and-palette-propagation-in-qt/
[2] https://github.com/Slicer/Slicer/blob/22f1c99f28d52f58070e31849d33a3a50e4501a6/Base/QTGUI/DesignerPlugins/qSlicerWidgetPlugin.cxx#L41