ENH: Control default settings in vesselview
With this change, we can control the default user settings in the
application without having to modify slicer's code. It works by checking if
a setting file already exists or not. If one already does not exists
already (for example the first time vesselview is started) then the
VesselViewDefaultSettings.ini is used. However if one exists already it's
left intact. It's thus transparent for the user as they can style change
the settings however they want.