Fix interactive segment tubes bug with seeds added during execution
Seeds added during the execution would be marked as processed even if they
had never went through CLI execution.
Fixing this implied quite a little bit of refactoring on how the CLI is
ran. Now the points are added to a list called queue. This queue is
transfered to a processing list and the todo markup when the CLI needs
to be executed.
The processing list helps to keep track of what point was processed in the
last CLI execution. When the CLI is finished, this list is emptied and
all the points it contained marked as processed.
The todo markup is here srictly for the CLI's benefit (passing the points
thought the CLI-node interface).

This commit also includes a bit of renaming:
 - All the "...ToDoMarkup..."  methods/function/variables were renamed
with to "...Todo...".
 - isRunning was renamed to isCLIRunning