Add testing
Add testing for vesselview. Right now only c++ testing are added but python
tests will follow.
To run the tests, run the ctest command in the
VesselView_SUPERBUILD_DIR/Slicer-build/Slicer-build/Applications.

To allow a simple configuration of the testing for the dashboard, one only
has to enable BUILD_TESTING on VesselView superbuild. Only the testing in
the application folder of VesselView will be ran (no Slicer test are ran).

In details, we need to pass a variable to the inner build of slicer since
the application folder is read as if it was slicer's own application folder
. Thus it can only "know" cmake variable that slicer would know.
We circumvent this using the VesselView build. This has two advantages:
 - We can communicate a special variable to Slicer inner build. Since
Slicer depends on VesselView, it will do a find_package on it, thus adding
all the variable VesselView exposes through its UseVesselView.cmake.
 -> This allow us to propagate the testing (or not) as we wish through the
 ${APPLICATION_NAME}_BUILD_TESTING variable.

 - (NOT TESTED YET) If we ever happen to have our own libraries, they will
know that they need to be tested. Note that this may be an issue as to
where we would need to run CTEST...