Move the tortuosity testing in the application folder
This is the occasion to refactor the cxx testing CMakeLists in
Applications. It is now testing the modules as well as the normal
application tests. It was refactored to use the slicer macro which makes
the code more readable and more slicer-like.

NOTE that the test fails for now ! Either the tortuosity module or the test
is wrong.