Recently loaded files - part 1
This is a bigger undertaking that I realzed and this is as good as any
stopping point.

Methods in the main window to access the recent files from an outside
class.
In the welcome screen, a list view was added to show the recent files. This
list leverages gutenberg for a file icon. Most importantly, the list view
is using a model/view scheme. The model and the proxy model (for hiding
irrelevant file type) have been implemented in the welcome screen.

The whole things is a bit of a mess and doesn't do much. Nothing happens
on click and the scrolling isn't correct. The next commit should take
care of that.