Recently loaded files - part 2
Actually load files that have been clicked on.

To make this work, I had to:
 - In the main window
Make sure that the signal that the settings have changed is correctly sent
even when we do "clear history". This meant cleaning up a little bit of
code to make it more uniform. For example, some methods were taking
the member as an argument. It's cleaner if they just use it without passing
it around.
Also make sure that the settings are wiped properly every time we write
them.

 - In the welcome screen
Add method in the model to allow for element removal. To do so, one needs
to re-implement the RemoveRows() methods.
Removed not-implemented method recentFile.
Add accessor to the filename given a row index in the proxy for the QML.
Add a filesToOpenOnLoad property in the welcome screen. This property is
used to stored which files will need to be loaded just before loading a
module. It is easier to handle lists in the c++ code than in the QML so
the property was made to be only easy calls in QML.

 - In the QML
The recentlyLoadedFilesListView was actually sitting on top of the
description rectangle instead of being under it. To correct this, the list
view was made a sibling (and not a child) of the rectangle. The same goes
for the open button.
The mouse area of the recentFilesListView's delegate handles all the
interaction and populates/depopulates the list of files to open on load.