ENH: Made UI changes in tube table widget.
Remove the mark as root column from the table. User have to select
tubes, selected tubes will be considered as roots while running tree
creation.
If a tube is a root and have a sub-tree then is Root column shows a menu
to manipulate the entire sub-tree.
Remove Show/Hide column from the table. Added it as option to affect all
the selected tubes. Added display properties option to change the
visualization of the selected tubes
Added an option to apply color to selected tube.